posting from fareas without .info

+++ ftntic.py
@@ -89,6 +89,19 @@ def read_tic(f):
     ticdata.setdefault(op.upper(), []).append(data)
   return ticdata
 
+def sz_crc32(fn):
+  checksum=0
+  f=open(fn, "rb")
+  fsize = 0
+  while(True):
+    z=f.read(262144)
+    fsize += len(z)
+    if not z:
+      break
+    checksum=zlib.crc32(z, checksum)
+  f.close()
+  return fsize, "%08X"%checksum
+
 def find_file(name, path):
   n1=os.path.join(path, name)
   if os.path.exists(n1):
@@ -102,15 +115,18 @@ def find_file(name, path):
   raise NoFile("no match for %s at %s"%(name, path))
 
 
-def import_tic(db, fullname, expect_addr=None, import_utime=None):
+def import_tic(db, fullname, expect_addr=None, import_utime=None, ticdata=None):
   # if "TO" is present
   #   get from links with matching from and to addresses and verify password
   # if "TO" is absent
   #   get to and password from links by from. if two rows are fetched - refuse tic
   #
   # in both cases refuse tic if no row fetched - tics are allowed for password links only
-  filepath, filename = os.path.split(fullname)
-  ticdata = read_tic(fullname)
+  if ticdata is None:
+    filepath, filename = os.path.split(fullname)
+    ticdata = read_tic(fullname)
+  else:
+    filepath = os.path.dirname(fullname)
 
   tic_src = get_optional(ticdata, "FROM")
   print ("from", tic_src)
@@ -120,47 +136,54 @@ def import_tic(db, fullname, expect_addr=None, import_utime=None):
   tic_dest = get_optional(ticdata, "TO")
   print ("to", tic_dest)
 
-  q="select l.address, l.my, l.authentication from links l"
-  q_args=[]
-  if tic_src:
-    src_id = ftnconfig.get_addr_id(db, db.FTN_domains["node"], tic_src)
-    q += (" and" if q_args else " where") + " address=$%d"%(len(q_args)+1)
-    q_args.append(src_id)
-  else:
+  if tic_src is None and tic_dest is None and skip_access_check:
+    print ("Importing non-FTN file")
     src_id = None
-
-  if tic_dest:
-    dest_id = ftnconfig.get_addr_id(db, db.FTN_domains["node"], tic_dest)
-    q += (" and" if q_args else " where") + " my=$%d"%(len(q_args)+1)
-    q_args.append(src_id)
-  else:
     dest_id = None
 
-  print (q)
-  print (q_args)
+  else:
 
-  possible_links = db.prepare(q)(*q_args)
-  if len(possible_links) > 1:
-    raise WrongTic("ambiguos link %s->%s"%(str(tic_src),str(tic_dest)))
+    q="select l.address, l.my, l.authentication from links l"
+    q_args=[]
+    if tic_src:
+      src_id = ftnconfig.get_addr_id(db, db.FTN_domains["node"], tic_src)
+      q += (" and" if q_args else " where") + " address=$%d"%(len(q_args)+1)
+      q_args.append(src_id)
+    else:
+      src_id = None
 
-  if len(possible_links) == 0:
-    raise WrongTic("no matching link %s->%s"%(str(tic_src),str(tic_dest)))
+    if tic_dest:
+      dest_id = ftnconfig.get_addr_id(db, db.FTN_domains["node"], tic_dest)
+      q += (" and" if q_args else " where") + " my=$%d"%(len(q_args)+1)
+      q_args.append(src_id)
+    else:
+      dest_id = None
 
-  src_id, dest_id, authinfo = possible_links[0]
-  pw = authinfo.find("RobotsPassword").text
+    print (q)
+    print (q_args)
 
-  print (src_id, dest_id, pw)
+    possible_links = db.prepare(q)(*q_args)
+    if len(possible_links) > 1:
+      raise WrongTic("ambiguos link %s->%s"%(str(tic_src),str(tic_dest)))
 
-  tic_passw = get_single(ticdata, "PW")
-  if tic_passw!=pw:
-    raise WrongTic("invalid password for %s"%tic_src)
+    if len(possible_links) == 0:
+      raise WrongTic("no matching link %s->%s"%(str(tic_src),str(tic_dest)))
+
+    src_id, dest_id, authinfo = possible_links[0]
+    pw = authinfo.find("RobotsPassword").text
+
+    print (src_id, dest_id, pw)
+
+    tic_passw = get_single(ticdata, "PW")
+    if tic_passw!=pw:
+      raise WrongTic("invalid password for %s"%tic_src)
 
   # source and destination verified, now try to find file
   # but before we should check if link can post to specified area
   area = get_single(ticdata, "AREA")
   print (area)
-  maypost = ftnaccess.may_post(db, tic_src, ("fileecho", area))
-  if not skip_access_check and not maypost:
+  if not skip_access_check:
+    maypost = ftnaccess.may_post(db, tic_src, ("fileecho", area))
     raise WrongTic("%s may not post to %s"%(tic_src, area))
 
   fname = os.path.split(get_single(ticdata, "FILE"))[1]
@@ -180,17 +203,9 @@ def import_tic(db, fullname, expect_addr=None, import_utime=None):
   if fsize is not None and os.path.getsize(ffullname)!=fsize:
     raise NoFile("file %s size != %d"%(ffullname, fsize))
 
-  checksum=0
-  f=open(ffullname, "rb")
-  fsize = 0
-  while(True):
-    z=f.read(262144)
-    fsize += len(z)
-    if not z:
-      break
-    checksum=zlib.crc32(z, checksum)
-  f.close()
-  if "%08X"%checksum != fcrc:
+  fsize, checksum = sz_crc32(ffullname)
+
+  if checksum != fcrc:
     raise NoFile("file %s crc32 %s != %s"%(ffullname, checksum, fcrc))
 
   print ("file matches")
@@ -207,20 +222,28 @@ def import_tic(db, fullname, expect_addr=None, import_utime=None):
     # select posting of same origin, area, filename, origin_record
     # if any has same filesize and hash - compare content and drop duplicate
 
-    tic_origin = get_single(ticdata, "ORIGIN")
-    with ftnimport.session(db) as sess:
-      tic_origin_id = sess.check_addr("node", tic_origin)
+    tic_origin = get_optional(ticdata, "ORIGIN")
+    if tic_origin:
+      with ftnimport.session(db) as sess:
+        tic_origin_id = sess.check_addr("node", tic_origin)
+    else:
+      tic_origin_id = None
+
     area_id = ftnconfig.get_addr_id(db, db.FTN_domains["fileecho"], area)
-    tic_originrec = get_first(ticdata, "PATH")
 
-    print("check if tic is first %s %d %s %s"%((tic_origin, area_id, fname, tic_originrec)))
+    try:
+      tic_originrec = get_first(ticdata, "PATH")
+      print("check if tic is first %s %d %s %s"%((tic_origin, area_id, fname, tic_originrec)))
 
-    for prev_f, prev_l, prev_h, prev_p in db.prepare("select f.id, f.length, f.sha512, p.id from files f inner join file_post p ON p.filedata=f.id "
-        "where p.origin=$1 and p.destination=$2 and p.filename=$3 and p.origin_record=$4")(tic_origin_id, area_id, fname, tic_originrec):
-      os.rename(ffullname, ffullname+".dup")
-      os.rename(fullname, fullname+".dup")
-      raise DupPost("similar posting %d, abandom"%prev_p, ffullname)
-      # tic with the same first record of PATH - the same posting
+      for prev_f, prev_l, prev_h, prev_p in db.prepare("select f.id, f.length, f.sha512, p.id from files f inner join file_post p ON p.filedata=f.id "
+          "where p.origin=$1 and p.destination=$2 and p.filename=$3 and p.origin_record=$4")(tic_origin_id, area_id, fname, tic_originrec):
+        os.rename(ffullname, ffullname+".dup")
+        os.rename(fullname, fullname+".dup")
+        raise DupPost("similar posting %d, abandom"%prev_p, ffullname)
+        # tic with the same first record of PATH - the same posting
+    except BadTic:
+      print ("PATH is missing, no dupe checking")
+      tic_originrec = None
 
     sha512 = hashlib.new("sha512")
     f=open(ffullname, "rb")
@@ -273,17 +296,17 @@ def import_tic(db, fullname, expect_addr=None, import_utime=None):
     else:
       utime = int(import_utime)
 
-    print ("current utime", utime)
-    while db.prepare("select id from file_post where post_time=$1").first(utime):
-      utime+=1
-    print ("insert with utime", utime)
+    print ("post_time=", utime)
 
     db.prepare("insert into file_post (filedata, origin, destination, recv_from, recv_as, recv_timestamp, origin_record, filename, other, post_time) "
-                    "values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)")\
+                    "values ($1, $2, $3, $4, $5, $6, $7, $8, $9, free_posttime($10))")\
       (f_id, tic_origin_id, area_id, src_id, dest_id, datetime.datetime.now(datetime.timezone.utc), tic_originrec, fname, json.dumps(ticdata), utime)
     print ("inserted successfully")
+    if not fullname.endswith(".faketic"):
+      print ("unlink", fullname)
+      os.unlink(fullname)
+    print ("unlink", ffullname)
     os.unlink(ffullname)
-    os.unlink(fullname)
 
   # <<< UNLOCK FILEECHOES POSTINGS
